Move 'Status' and other c-structs out to FFI module
